Split off part of identity.go into identity_repr.go
